5396: rename files to card list page
