+++ src/stories/Library/card-list-page/CardListPage.tsx
@@ -5,9 +5,9 @@ import { SearchResultTitle } from "./SearchResultTitle";
 import { SearchResultZero } from "./SearchResultZero";
 import FacetLine from "./FacetLine";
 import FacetLineSelected from "./FacetLineSelectedTerms";
-import data from "./SearchResultPageData";
+import data from "./CardListPageData";
 
-export type SearchResultPageProps = {
+export type CardListPageProps = {
   title: string;
   currentResults: number;
   totalResults: number;
@@ -20,14 +20,14 @@ const SearchResultList = data.searchResult.map((item, i) => {
   return <CardListItem {...item} tintIndex={i} />;
 });
 
-export const SearchResultPage = ({
+export const CardListPage = ({
   title,
   linkName,
   linkTotalResults,
   currentResults,
   totalResults,
   zeroResult,
-}: SearchResultPageProps) => {
+}: CardListPageProps) => {
   return (
     <div className="card-list-page">
       <SearchResultTitle
